fix float to int

+++ lanes.py
@@ -36,9 +36,9 @@ def coordinates(image, line):
 	
 	y1 = image.shape[0]
 	# keeping y2 restricted to 1/2 of image length
-	y2 = y1 * (1/2)
-	x1 = (y1 - c) / m
-	x2 = (y2 - c) / m
+	y2 = int(y1 * (1/2))
+	x1 = int((y1 - c) / m)
+	x2 = int((y2 - c) / m)
 	
 	return np.array([x1, y1, x2, y2])
